@@ -10,22 +10,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.12",
-    "@angular/common": "~13.3.12",
-    "@angular/compiler": "~13.3.12",
-    "@angular/core": "~13.3.12",
-    "@angular/forms": "~13.3.12",
-    "@angular/platform-browser": "~13.3.12",
-    "@angular/platform-browser-dynamic": "~13.3.12",
-    "@angular/router": "~13.3.12",
+    "@angular/animations": "^14.2.12",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "rxjs": "^7.4.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.10",
-    "@angular/cli": "~13.3.10",
-    "@angular/compiler-cli": "~13.3.12",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~3.8.0",
@@ -34,7 +34,7 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "ng-packagr": "^13.3.1",
+    "ng-packagr": "^14.2.2",
     "typescript": "~4.6.4"
   }
 }
